add latitude and longitude recongnizer
Add latitude and logitude using the tap recognizer to return the
position of where the user tap on the map.

+++ VirtualTourist/Model/Pin.swift
@@ -9,8 +9,8 @@
 import Foundation
 
 struct Pin {
-    static var longitude: Float? = nil
-    static var latitude: Float? = nil
+    static var longitude: Double? = nil
+    static var latitude: Double? = nil
 }
 
 let shareInstance = Pin()
